Split post 1415 oversized slices
