+++ app/src/test/java/com/pocketshell/app/composer/PromptComposerViewModelTest.kt
@@ -256,139 +256,6 @@ class PromptComposerViewModelTest {
     // -- FSM transitions ----------------------------------------------------
 
     @Test
-    fun initialStateIsIdleWithEmptyDraft() {
-        val vm = newVm()
-        val state = vm.uiState.value
-        assertEquals(RecordingState.Idle, state.recording)
-        assertEquals("", state.draft)
-        assertNull(state.error)
-    }
-
-    @Test
-    fun diagnosticsRecordAttachmentSeedAndRemove() {
-        val diagnostics = installRecordingDiagnosticSink()
-        try {
-            val vm = newVm()
-
-            vm.seedAttachment("/home/alexey/private/report.txt")
-            vm.removeAttachment("/home/alexey/private/report.txt")
-
-            assertEquals(
-                listOf("attachment_seeded_from_share", "attachment_remove"),
-                diagnostics.events.map { it.name },
-            )
-            assertEquals(1, diagnostics.events[0].fields["attachmentCount"])
-            assertEquals(1, diagnostics.events[1].fields["beforeCount"])
-            assertEquals(0, diagnostics.events[1].fields["afterCount"])
-            assertFalse(diagnostics.events.any { event ->
-                event.fields.values.any { it == "/home/alexey/private/report.txt" }
-            })
-        } finally {
-            diagnostics.close()
-        }
-    }
-
-    @Test
-    fun seedDraftPromptReplacesEmptyDraft() {
-        val vm = newVm()
-        vm.seedDraftPrompt("Apply the PocketShell review at ~/inbox/pocketshell/reviews/README.md-20260614-025147.yaml")
-        assertEquals(
-            "Apply the PocketShell review at ~/inbox/pocketshell/reviews/README.md-20260614-025147.yaml",
-            vm.uiState.value.draft,
-        )
-    }
-
-    @Test
-    fun seedDraftPromptAppendsToExistingDraftOnNewLine() {
-        val vm = newVm()
-        vm.onDraftChange("look at this")
-        vm.seedDraftPrompt("Apply the PocketShell review at /tmp/r.yaml")
-        assertEquals(
-            "look at this\nApply the PocketShell review at /tmp/r.yaml",
-            vm.uiState.value.draft,
-        )
-    }
-
-    @Test
-    fun seedDraftPromptIgnoresBlank() {
-        val vm = newVm()
-        vm.onDraftChange("keep me")
-        vm.seedDraftPrompt("   ")
-        assertEquals("keep me", vm.uiState.value.draft)
-    }
-
-    @Test
-    fun seedDraftPromptRecordsDiagnostic() {
-        val diagnostics = installRecordingDiagnosticSink()
-        try {
-            val vm = newVm()
-            vm.seedDraftPrompt("Apply the PocketShell review at /tmp/r.yaml")
-            assertEquals(
-                listOf("review_prompt_seeded_into_composer"),
-                diagnostics.events.map { it.name },
-            )
-        } finally {
-            diagnostics.close()
-        }
-    }
-
-    // -- #770: tap a rendered engine command -> composer pre-filled ----------
-
-    @Test
-    fun prefillEngineCommandSetsTheDraftToTheCommand() {
-        val vm = newVm()
-        vm.prefillEngineCommand("/clear")
-        assertEquals("/clear", vm.uiState.value.draft)
-    }
-
-    @Test
-    fun prefillEngineCommandReplacesAnExistingLeadingSlashToken() {
-        val vm = newVm()
-        vm.onDraftChange("/comp")
-        vm.prefillEngineCommand("/clear")
-        assertEquals("/clear", vm.uiState.value.draft)
-    }
-
-    @Test
-    fun prefillEngineCommandPreservesTrailingTextAfterTheSlashToken() {
-        val vm = newVm()
-        vm.onDraftChange("/old keep this part")
-        vm.prefillEngineCommand("/clear")
-        assertEquals("/clear keep this part", vm.uiState.value.draft)
-    }
-
-    @Test
-    fun prefillEngineCommandPrependsWhenDraftHasNoLeadingSlash() {
-        val vm = newVm()
-        vm.onDraftChange("already typed")
-        vm.prefillEngineCommand("/clear")
-        assertEquals("/clearalready typed", vm.uiState.value.draft)
-    }
-
-    @Test
-    fun prefillEngineCommandIgnoresBlank() {
-        val vm = newVm()
-        vm.onDraftChange("keep me")
-        vm.prefillEngineCommand("   ")
-        assertEquals("keep me", vm.uiState.value.draft)
-    }
-
-    @Test
-    fun prefillEngineCommandRecordsDiagnostic() {
-        val diagnostics = installRecordingDiagnosticSink()
-        try {
-            val vm = newVm()
-            vm.prefillEngineCommand("/clear")
-            assertEquals(
-                listOf("engine_command_tapped_into_composer"),
-                diagnostics.events.map { it.name },
-            )
-        } finally {
-            diagnostics.close()
-        }
-    }
-
-    @Test
     fun micTapInIdleStartsRecording() = runTest {
         val mic = FakeMicCapture()
         val vm = newVm(mic = mic, samplerDispatcher = StandardTestDispatcher(testScheduler))
