+++ app/src/test/java/com/pocketshell/app/tmux/TmuxSessionViewModelTest.kt
@@ -2880,1409 +2880,6 @@ class TmuxSessionViewModelTest : TmuxSessionViewModelTestBase() {
     }
 
     @Test
-    fun networkChangeLifecycleHookEntersReconnectingWithoutConnectionError() = runTest(scheduler) {
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setAutoReconnectDelaysForTest(listOf(60_000L))
-        val client = FakeTmuxClient()
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = client,
-        )
-        runCurrent()
-
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(
-            networkChange(reason = "validated-default-network-changed"),
-        )
-        runCurrent()
-
-        assertEquals(
-            "network change should not wait for the tmux reader EOF",
-            0,
-            connector.connectCount,
-        )
-        val status = vm.connectionStatus.value
-        assertTrue(
-            "network change must show reconnect-in-progress, not a connection error; got $status",
-            status is TmuxSessionViewModel.ConnectionStatus.Reconnecting,
-        )
-        assertTrue(
-            "reconnect reason should avoid misleading manual retry wording",
-            "Tap Reconnect" !in (status as TmuxSessionViewModel.ConnectionStatus.Reconnecting).reason,
-        )
-        assertEquals(TmuxConnectTrigger.NetworkReconnect, vm.latestRestoreIntentSnapshot()?.trigger)
-        assertTrue("manual reconnect remains available during proactive reconnect", vm.canReconnect.value)
-        assertTrue(
-            "proactive reconnect should remove the stale active client from the registry",
-            registry.clients.value.isEmpty(),
-        )
-    }
-
-    @Test
-    fun sameNetworkHandleWithDifferentTransportsDoesNotEnterReconnecting() = runTest(scheduler) {
-        val diagnostics = installRecordingDiagnosticSink()
-        TMUX_CONNECT_ATTEMPTS.set(0)
-        try {
-            val registry = ActiveTmuxClients()
-            val connector = QueueLeaseConnector(FakeSshSession())
-            val vm = newVm(
-                registry = registry,
-                sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-            )
-            vm.setAutoReconnectDelaysForTest(listOf(0L))
-            val client = FakeTmuxClient()
-            vm.replaceClientForTest(
-                hostId = 7L,
-                hostName = "alpha",
-                host = "alpha.example",
-                port = 22,
-                user = "alex",
-                keyPath = "/keys/a",
-                sessionName = "work",
-                client = client,
-            )
-            runCurrent()
-
-            registry.lifecycleHooksSnapshot().single().onNetworkChanged(
-                networkChange(
-                    previous = TerminalNetworkSnapshot.Validated("same-handle", setOf("WIFI")),
-                    current = TerminalNetworkSnapshot.Validated("same-handle", setOf("VPN", "WIFI")),
-                    previousValidated = TerminalNetworkSnapshot.Validated("same-handle", setOf("WIFI")),
-                    reason = "same-handle-transport-metadata",
-                ),
-            )
-            advanceUntilIdle()
-
-            assertTrue(
-                "same network handle with changed transports is metadata churn, got ${vm.connectionStatus.value}",
-                vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected,
-            )
-            assertEquals(0, connector.connectCount)
-            assertEquals(0, TMUX_CONNECT_ATTEMPTS.get())
-            assertSame(client, registry.clients.value[7L]?.client)
-            val skip = diagnostics.eventsNamed("network_reconnect_skip").single()
-            assertEquals("no_real_validated_handoff", skip.fields["cause"])
-            assertEquals("network_identity_unchanged", skip.fields["classification"])
-            assertEquals(false, skip.fields["reconnect"])
-            assertEquals(false, skip.fields["realValidatedIdentityChange"])
-            assertEquals("same-handle", skip.fields["currentNetworkHandle"])
-            assertEquals("VPN,WIFI", skip.fields["currentTransports"])
-            val trail = diagnostics.eventsNamed("cause_trail")
-                .single { it.fields["stage"] == "network_reconnect_decision" }
-            assertEquals("suppress", trail.fields["outcome"])
-            assertEquals("no_real_validated_handoff", trail.fields["cause"])
-            assertEquals("network_identity_unchanged", trail.fields["classification"])
-            assertEquals(7L, trail.fields["hostId"])
-        } finally {
-            diagnostics.close()
-        }
-    }
-
-    @Test
-    fun differentNetworkHandleStillEntersReconnecting() = runTest(scheduler) {
-        val diagnostics = installRecordingDiagnosticSink()
-        try {
-            val registry = ActiveTmuxClients()
-            val connector = QueueLeaseConnector(FakeSshSession())
-            val vm = newVm(
-                registry = registry,
-                sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-            )
-            vm.setAutoReconnectDelaysForTest(listOf(60_000L))
-            vm.replaceClientForTest(
-                hostId = 7L,
-                hostName = "alpha",
-                host = "alpha.example",
-                port = 22,
-                user = "alex",
-                keyPath = "/keys/a",
-                sessionName = "work",
-                client = FakeTmuxClient(),
-            )
-            runCurrent()
-
-            registry.lifecycleHooksSnapshot().single().onNetworkChanged(
-                networkChange(
-                    previous = TerminalNetworkSnapshot.Validated("wifi", setOf("WIFI")),
-                    current = TerminalNetworkSnapshot.Validated("cell", setOf("CELLULAR")),
-                    previousValidated = TerminalNetworkSnapshot.Validated("wifi", setOf("WIFI")),
-                    reason = "different-network-handle",
-                ),
-            )
-            runCurrent()
-
-            assertTrue(vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Reconnecting)
-            assertEquals(TmuxConnectTrigger.NetworkReconnect, vm.latestRestoreIntentSnapshot()?.trigger)
-            assertEquals(0, connector.connectCount)
-            assertTrue("network reconnect removes stale active client", registry.clients.value.isEmpty())
-            val start = diagnostics.eventsNamed("network_reconnect_start").single()
-            assertEquals("proactive_network_handoff", start.fields["classification"])
-            assertEquals(true, start.fields["reconnect"])
-            assertEquals(true, start.fields["realValidatedIdentityChange"])
-            assertEquals("wifi", start.fields["previousValidatedNetworkHandle"])
-            assertEquals("cell", start.fields["currentNetworkHandle"])
-            val trail = diagnostics.eventsNamed("cause_trail")
-                .single { it.fields["stage"] == "network_reconnect_decision" }
-            assertEquals("schedule_reconnect", trail.fields["outcome"])
-            assertEquals("proactive_network_handoff", trail.fields["cause"])
-            assertEquals("proactive_network_handoff", trail.fields["classification"])
-            assertEquals(7L, trail.fields["hostId"])
-        } finally {
-            diagnostics.close()
-        }
-    }
-
-    @Test
-    fun restoredSameNetworkHintDuringActiveTerminalOutputDoesNotShowReconnect() = runTest(scheduler) {
-        TMUX_CONNECT_ATTEMPTS.set(0)
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setAutoReconnectDelaysForTest(listOf(0L, 0L, 0L, 0L))
-        val client = FakeTmuxClient()
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = client,
-        )
-        vm.applyParsedPanesForTest(
-            listOf(
-                TmuxSessionViewModel.ParsedPane(
-                    "%0",
-                    "@0",
-                    "\$0",
-                    "work",
-                    paneIndex = 0,
-                    sessionName = "work",
-                ),
-            ),
-        )
-        advanceUntilIdle()
-
-        val state = vm.panes.value.single().terminalState
-        val slowSideChannelCollector = launch {
-            state.output.collect {
-                delay(60_000)
-            }
-        }
-        runCurrent()
-
-        try {
-            client.emittedEvents.emit(ControlEvent.Output("%0", issue576BackpressureOutputChunks().first()))
-            runCurrent()
-
-            val hook = registry.lifecycleHooksSnapshot().single()
-            repeat(4) { index ->
-                hook.onNetworkChanged(
-                    networkChange(
-                        previous = TerminalNetworkSnapshot.NoValidatedNetwork,
-                        current = TerminalNetworkSnapshot.Validated("wifi"),
-                        previousValidated = TerminalNetworkSnapshot.Validated("wifi"),
-                        reason = "validated-network-during-output-$index",
-                    ),
-                )
-                runCurrent()
-            }
-
-            assertTrue(
-                "active terminal output proves the tmux stream is alive; network hints must not show reconnect",
-                vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected,
-            )
-            assertEquals(
-                "network hints during active output must not enqueue reconnect attempts",
-                0,
-                connector.connectCount,
-            )
-            assertEquals(
-                "network hints during active output must not increment the reconnect counter",
-                0,
-                TMUX_CONNECT_ATTEMPTS.get(),
-            )
-            assertFalse(
-                "terminal-side backpressure must not flip the tmux disconnected signal",
-                client.disconnectedSignal.value,
-            )
-        } finally {
-            slowSideChannelCollector.cancel()
-        }
-    }
-
-    @Test
-    fun outputForOnlyTerminalOutputKeepsNetworkHintFromShowingReconnect() = runTest(scheduler) {
-        TMUX_CONNECT_ATTEMPTS.set(0)
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setAutoReconnectDelaysForTest(listOf(0L, 0L))
-        val client = FakeTmuxClient()
-        client.decoupleOutputForFromEvents = true
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = client,
-        )
-        vm.applyParsedPanesForTest(
-            listOf(
-                TmuxSessionViewModel.ParsedPane(
-                    "%0",
-                    "@0",
-                    "\$0",
-                    "work",
-                    paneIndex = 0,
-                    sessionName = "work",
-                ),
-            ),
-        )
-        advanceUntilIdle()
-
-        client.emittedPaneOutputs.emit(ControlEvent.Output("%0", "visible via outputFor".toByteArray()))
-        runCurrent()
-
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(
-            networkChange(
-                previous = TerminalNetworkSnapshot.NoValidatedNetwork,
-                current = TerminalNetworkSnapshot.Validated("wifi"),
-                previousValidated = TerminalNetworkSnapshot.Validated("wifi"),
-                reason = "same-network-hint-after-outputFor-only-output",
-            ),
-        )
-        runCurrent()
-
-        assertTrue(
-            "visible output delivered without client.events %output must still suppress same-network reconnect",
-            vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected,
-        )
-        assertEquals(0, connector.connectCount)
-        assertEquals(0, TMUX_CONNECT_ATTEMPTS.get())
-    }
-
-    @Test
-    fun firstValidatedNetworkHintDoesNotReconnectIdleStableTerminal() = runTest(scheduler) {
-        TMUX_CONNECT_ATTEMPTS.set(0)
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setAutoReconnectDelaysForTest(listOf(0L, 0L))
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = FakeTmuxClient(),
-        )
-        advanceUntilIdle()
-
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(
-            networkChange(
-                previous = TerminalNetworkSnapshot.NoValidatedNetwork,
-                current = TerminalNetworkSnapshot.Validated("wifi"),
-                previousValidated = null,
-                reason = "first-validated-network",
-            ),
-        )
-        advanceUntilIdle()
-
-        assertTrue(
-            "first validated callback is startup state discovery, not proof that SSH died",
-            vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected,
-        )
-        assertEquals(0, connector.connectCount)
-        assertEquals(0, TMUX_CONNECT_ATTEMPTS.get())
-    }
-
-    @Test
-    fun realNetworkIdentityChangeDuringActiveTerminalOutputStillReconnects() = runTest(scheduler) {
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setAutoReconnectDelaysForTest(listOf(60_000L))
-        val client = FakeTmuxClient()
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = client,
-        )
-        runCurrent()
-        client.emittedEvents.emit(ControlEvent.Output("%0", "still streaming".toByteArray()))
-        runCurrent()
-
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(
-            networkChange(
-                previous = TerminalNetworkSnapshot.Validated("wifi"),
-                current = TerminalNetworkSnapshot.Validated("cell"),
-                previousValidated = TerminalNetworkSnapshot.Validated("wifi"),
-                reason = "wifi-cellular-handoff",
-            ),
-        )
-        runCurrent()
-
-        assertEquals(
-            "real validated identity change should still use proactive reconnect despite recent output",
-            TmuxConnectTrigger.NetworkReconnect,
-            vm.latestRestoreIntentSnapshot()?.trigger,
-        )
-        assertTrue(vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Reconnecting)
-        assertEquals(0, connector.connectCount)
-    }
-
-    // ---- Issue #981: network handoff rides through while the transport is proven alive ----
-
-    @Test
-    fun issue981ValidatedNetworkFlipDoesNotTearDownTransportProvenAliveLink() = runTest(scheduler) {
-        // Issue #981 reproduce-first (red on base): a REAL validated default-network
-        // identity flip (WIFI→CELLULAR) arrives while the live SSH transport is
-        // provably alive (its keepalive saw inbound bytes within the ride-through
-        // window — the #974 stable-wifi case where -CC traffic keeps the link warm).
-        // The reactive handoff path must NOT tear down + redial the healthy socket.
-        // On base (no liveness gate) this asserts Reconnecting + 1 connect and FAILS;
-        // with the fix it rides through (Connected, zero connects).
-        TMUX_CONNECT_ATTEMPTS.set(0)
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setAutoReconnectDelaysForTest(listOf(0L))
-        // A live session whose transport keepalive proves the link is alive.
-        val session = FakeSshSession().apply { transportProvenAlive = true }
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = FakeTmuxClient(),
-            session = session,
-        )
-        advanceUntilIdle()
-        assertTrue(
-            "precondition: the link is the proven-alive transport the gate must ride through",
-            vm.isTransportKeepAliveProvenAliveRecentlyForTest(),
-        )
-
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(
-            networkChange(
-                previous = TerminalNetworkSnapshot.Validated("wifi"),
-                current = TerminalNetworkSnapshot.Validated("cell"),
-                previousValidated = TerminalNetworkSnapshot.Validated("wifi"),
-                reason = "transient-wifi-cellular-flip-on-stable-wifi",
-            ),
-        )
-        advanceUntilIdle()
-
-        assertTrue(
-            "a validated network flip while the transport is PROVEN ALIVE must ride through, " +
-                "not tear down the healthy socket (#981 / #974 stable-wifi drop)",
-            vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected,
-        )
-        assertEquals(
-            "no redial may be scheduled while the old transport is provably alive",
-            0,
-            connector.connectCount,
-        )
-        assertEquals(0, TMUX_CONNECT_ATTEMPTS.get())
-        assertNotEquals(
-            "the proactive-handoff redial intent must NOT fire on a proven-alive link",
-            TmuxConnectTrigger.NetworkReconnect,
-            vm.latestRestoreIntentSnapshot()?.trigger,
-        )
-    }
-
-    @Test
-    fun issue981ValidatedNetworkFlipStillReconnectsWhenTransportIsDead() = runTest(scheduler) {
-        // Issue #981 class-coverage (G2): the gate must NOT mask a GENUINE handoff.
-        // Same WIFI→CELLULAR flip, but the transport keepalive has aged out (the old
-        // socket is really dead) → the path must STILL reconnect, exactly as before.
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setAutoReconnectDelaysForTest(listOf(60_000L))
-        // A live session whose transport keepalive has stopped proving liveness.
-        val session = FakeSshSession().apply { transportProvenAlive = false }
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = FakeTmuxClient(),
-            session = session,
-        )
-        runCurrent()
-        assertFalse(
-            "precondition: the dead transport must NOT report proven-alive",
-            vm.isTransportKeepAliveProvenAliveRecentlyForTest(),
-        )
-
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(
-            networkChange(
-                previous = TerminalNetworkSnapshot.Validated("wifi"),
-                current = TerminalNetworkSnapshot.Validated("cell"),
-                previousValidated = TerminalNetworkSnapshot.Validated("wifi"),
-                reason = "genuine-wifi-cellular-handoff-dead-link",
-            ),
-        )
-        runCurrent()
-
-        assertEquals(
-            "a genuine handoff on a DEAD link must still proactively reconnect (gate must not mask recovery)",
-            TmuxConnectTrigger.NetworkReconnect,
-            vm.latestRestoreIntentSnapshot()?.trigger,
-        )
-        assertTrue(vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Reconnecting)
-    }
-
-    @Test
-    fun issue981SameIdentityReassocStillSuppressedRegardlessOfTransportLiveness() = runTest(scheduler) {
-        // Issue #981 class-coverage: a same-identity reassoc (#875 pure-{WIFI} roam)
-        // is still suppressed by hasSameNetworkIdentityAs BEFORE the liveness gate,
-        // so the new gate does not change that path — proven-alive or not.
-        TMUX_CONNECT_ATTEMPTS.set(0)
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setAutoReconnectDelaysForTest(listOf(0L))
-        val session = FakeSshSession().apply { transportProvenAlive = false }
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = FakeTmuxClient(),
-            session = session,
-        )
-        advanceUntilIdle()
-
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(
-            networkChange(
-                previous = TerminalNetworkSnapshot.Validated("wifi"),
-                current = TerminalNetworkSnapshot.Validated("wifi"),
-                previousValidated = TerminalNetworkSnapshot.Validated("wifi"),
-                reason = "same-identity-reassoc",
-            ),
-        )
-        advanceUntilIdle()
-
-        assertTrue(
-            "a same-identity reassoc is suppressed upstream (#875) and never reaches the redial",
-            vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected,
-        )
-        assertEquals(0, connector.connectCount)
-        assertEquals(0, TMUX_CONNECT_ATTEMPTS.get())
-    }
-
-    @Test
-    fun issue981ReducerSuppressesProvenAliveButSchedulesDeadLinkFlip() = runTest(scheduler) {
-        // Issue #981 reducer-level proof: the decision classifier itself returns
-        // SuppressNetworkTransportProvenAlive when proven alive and
-        // ScheduleNetworkReconnect when the link is dead, using the explicit seam.
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setAutoReconnectDelaysForTest(listOf(60_000L))
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = FakeTmuxClient(),
-        )
-        runCurrent()
-
-        // Proven alive → ride through.
-        vm.forceTransportProvenAliveForTest = true
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(
-            networkChange(reason = "flip-while-proven-alive"),
-        )
-        runCurrent()
-        assertTrue(
-            "proven-alive flip rides through (no Reconnecting band)",
-            vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected,
-        )
-        assertEquals(0, connector.connectCount)
-
-        // Aged out (dead) → reconnect.
-        vm.forceTransportProvenAliveForTest = false
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(
-            networkChange(reason = "flip-while-dead", sequence = 2L),
-        )
-        runCurrent()
-        assertEquals(
-            "once the keepalive is dead the same flip must schedule the proactive reconnect",
-            TmuxConnectTrigger.NetworkReconnect,
-            vm.latestRestoreIntentSnapshot()?.trigger,
-        )
-        assertTrue(vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Reconnecting)
-
-        vm.forceTransportProvenAliveForTest = null
-    }
-
-    // ---- Issue #997: bare network LOSS → hold (no churn) → RESTORE → fast reconnect.
-    // Pre-#997 a clean loss produced NO proactive change at all (the detector
-    // swallowed it), so a loss-suspended session never fast-recovered — it waited
-    // ~90s for the keepalive ride-through. The reducer arms here are the
-    // ViewModel half of the fix: a NetworkLost holds the lease + surfaces the calm
-    // band without churning; a NetworkRestored drives `scheduleAutoReconnect` even
-    // from a non-Connected state and bypasses the proven-alive gate.
-
-    @Test
-    fun issue997BareNetworkLossHoldsTheLeaseWithoutChurningOrTearingDown() = runTest(scheduler) {
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setAutoReconnectDelaysForTest(listOf(60_000L))
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = FakeTmuxClient(),
-        )
-        runCurrent()
-        assertTrue(vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected)
-
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(networkLoss())
-        runCurrent()
-
-        // The calm "reconnecting" band is surfaced (UI not left on a dead-but-live
-        // session) WITHOUT launching a redial loop and WITHOUT tearing the lease.
-        assertTrue(
-            "a bare loss surfaces the calm Reconnecting band immediately",
-            vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Reconnecting,
-        )
-        assertEquals("no redial during the loss window — lease is held", 0, connector.connectCount)
-
-        // No churn: even after time passes nothing redials (no ladder running).
-        advanceTimeBy(60_000L)
-        advanceUntilIdle()
-        assertEquals(0, connector.connectCount)
-    }
-
-    @Test
-    fun issue997NetworkRestoreDrivesFastReconnectFromLossSuspendedState() = runTest(scheduler) {
-        TMUX_CONNECT_ATTEMPTS.set(0)
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val reconnectClient = FakeTmuxClient().withSinglePane("work", "%1")
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setTmuxClientFactoryForTest { _, sessionName, _ ->
-            assertEquals("work", sessionName)
-            reconnectClient
-        }
-        vm.setAutoReconnectDelaysForTest(listOf(0L))
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = FakeTmuxClient(),
-        )
-        runCurrent()
-
-        // Issue #1042: this case proves a restore CAN redial FROM the loss-suspended
-        // (non-Connected) state — the #997 gap. Under #1042 the restore is now
-        // liveness-first, so inject a GENUINELY-DEAD transport (keepalive not proven
-        // alive by default + the bounded restore probe DEAD) so the liveness gate
-        // falls through to the fresh-lease redial being asserted here.
-        vm.forceLivenessProbeDeadForTest = true
-
-        // Loss: hold + flip to Reconnecting (the loss-suspended state).
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(networkLoss())
-        runCurrent()
-        assertTrue(
-            "loss leaves the session in the loss-suspended Reconnecting state",
-            vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Reconnecting,
-        )
-        assertEquals(0, connector.connectCount)
-
-        // Restore: must drive a FAST reconnect even though status is NOT Connected
-        // (the #997 gap — the Connected-only ScheduleNetworkReconnect path would
-        // have Ignored this).
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(networkRestore())
-        advanceUntilIdle()
-
-        assertEquals(
-            "restore must redial via the auto-reconnect ladder with the network trigger",
-            TmuxConnectTrigger.NetworkReconnect,
-            vm.latestRestoreIntentSnapshot()?.trigger,
-        )
-        assertEquals("exactly one fresh-lease redial on restore", 1, connector.connectCount)
-        assertSame(reconnectClient, registry.clients.value[7L]?.client)
-    }
-
-    @Test
-    fun issue1042NetworkRestoreRidesThroughWhenTransportProvenAlive() = runTest(scheduler) {
-        // Issue #1042 (cause #1) REFINES the pre-#1042 #997 restore behaviour. The old
-        // assumption — "a loss means the socket is DEAD, so always redial" — fired a
-        // spurious fresh-lease reconnect on every brief cellular dip even when the TCP
-        // socket survived. The restore is now LIVENESS-FIRST: when the existing
-        // transport is proven alive it RIDES THROUGH with NO redial. Pin
-        // proven-alive=true and confirm the restore does NOT redial and the session
-        // returns to Connected. (The genuinely-dead case still redials —
-        // issue997NetworkRestoreDrivesFastReconnectFromLossSuspendedState guards it.)
-        TMUX_CONNECT_ATTEMPTS.set(0)
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val reconnectClient = FakeTmuxClient().withSinglePane("work", "%1")
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setTmuxClientFactoryForTest { _, _, _ -> reconnectClient }
-        vm.setAutoReconnectDelaysForTest(listOf(0L))
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = FakeTmuxClient(),
-        )
-        runCurrent()
-
-        vm.forceTransportProvenAliveForTest = true
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(networkLoss())
-        runCurrent()
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(networkRestore())
-        advanceUntilIdle()
-
-        assertEquals(
-            "a restore over a PROVEN-ALIVE transport must NOT redial (rides through) — #1042",
-            0,
-            connector.connectCount,
-        )
-        assertNull(
-            "a ride-through restore must not schedule a network-reconnect redial",
-            vm.latestRestoreIntentSnapshot(),
-        )
-        assertTrue(
-            "the ride-through clears the loss-hold band back to Connected",
-            vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected,
-        )
-
-        vm.forceTransportProvenAliveForTest = null
-    }
-
-    @Test
-    fun issue1042NetworkRestoreRidesThroughWhenBoundedProbeAnswers() = runTest(scheduler) {
-        // Issue #1042 (cause #1) ARM 2 — the headline symptom path: a quiet/idle
-        // cellular session whose transport keepalive has AGED OUT of its ride-through
-        // window (so the proven-alive fast gate does NOT fire) but whose TCP socket is
-        // still alive. The restore must then issue ONE bounded control-channel probe
-        // and, when it ANSWERS, RIDE THROUGH with no redial — NOT redial. This arm is
-        // distinct from arm 1 (keepalive proven alive) and arm 3 (neither answers ->
-        // fresh-lease redial). Asserted via the `cause="probe_answered"` device trail
-        // so it cannot be confused with the proven-alive arm.
-        TMUX_CONNECT_ATTEMPTS.set(0)
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val reconnectClient = FakeTmuxClient().withSinglePane("work", "%1")
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setTmuxClientFactoryForTest { _, _, _ -> reconnectClient }
-        vm.setAutoReconnectDelaysForTest(listOf(0L))
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = FakeTmuxClient(),
-        )
-        runCurrent()
-
-        // Keepalive NOT proven alive (the fast gate must MISS so the bounded-probe arm
-        // runs) but the bounded probe ANSWERS (healthy attached client, probe-dead seam
-        // off) — the surviving-but-quiet-link case.
-        vm.forceTransportProvenAliveForTest = false
-        vm.forceLivenessProbeDeadForTest = false
-
-        val diagnostics = installRecordingDiagnosticSink()
-        try {
-            registry.lifecycleHooksSnapshot().single().onNetworkChanged(networkLoss())
-            runCurrent()
-            registry.lifecycleHooksSnapshot().single().onNetworkChanged(networkRestore())
-            advanceUntilIdle()
-
-            assertEquals(
-                "a restore whose bounded probe ANSWERS must NOT redial (rides through) — #1042 arm 2",
-                0,
-                connector.connectCount,
-            )
-            assertNull(
-                "a probe-answered ride-through must not schedule a network-reconnect redial",
-                vm.latestRestoreIntentSnapshot(),
-            )
-            assertTrue(
-                "the probe-answered ride-through clears the loss-hold band back to Connected",
-                vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected,
-            )
-            assertTrue(
-                "no redial diagnostic may fire on the probe-answered arm; events=" +
-                    diagnostics.events.map { it.name },
-                diagnostics.eventsNamed("network_restore_reconnect_start").isEmpty() &&
-                    diagnostics.eventsNamed("reconnect_start").isEmpty(),
-            )
-            val rideThrough = diagnostics.eventsNamed("network_restore_ride_through")
-            assertTrue(
-                "expected a network_restore_ride_through (proves arm 2 fired, not a vacuous " +
-                    "pass); events=${diagnostics.events.map { it.name }}",
-                rideThrough.isNotEmpty(),
-            )
-            assertEquals(
-                "the ride-through must be attributed to the bounded probe answering " +
-                    "(distinguishes arm 2 from the proven-alive arm 1)",
-                "probe_answered",
-                rideThrough.single().fields["cause"],
-            )
-        } finally {
-            diagnostics.close()
-            vm.forceTransportProvenAliveForTest = null
-        }
-    }
-
-    @Test
-    fun issue1193NetworkRestoreFreshKeepaliveButDeadSocketRedialsViaProbeNotPassiveRideThrough() =
-        runTest(scheduler) {
-            // Issue #1193 REPRODUCE-FIRST (D33/G10): the maintainer's 2026-07-02
-            // cellular spurious drop. On a WiFi→cellular restore the OLD socket is
-            // silently dead (the new radio's IP/NAT invalidates the 4-tuple) but the
-            // PASSIVE keepalive TIMESTAMP is still fresh (< 90s). Pre-#1193 branch 1 of
-            // scheduleNetworkReconnectOnRestore rode through on that timestamp ALONE
-            // (isTransportKeepAliveProvenAliveRecently — a pure clock comparison, NO
-            // round-trip) → committed to the dead transport, flipped back to Live, and
-            // the `-CC` reader threw ~157ms later → passive_disconnect + reconnect
-            // churn. This pins BOTH the fresh keepalive (forceTransportProvenAliveForTest
-            // = true, the maintainer's EXACT state) AND a genuinely dead socket
-            // (forceLivenessProbeDeadForTest = true — the bounded probe won't answer).
-            //
-            // RED (branch 1 present): records a network_restore_ride_through
-            //   cause=transport_proven_alive and does NOT redial (connectCount == 0) —
-            //   the app commits to the dead transport.
-            // GREEN (#1193, branch 1 deleted → all restores go through the bounded
-            //   probe): the probe requires an answered round-trip, finds the socket
-            //   dead, and cleanly redials — network_restore_reconnect_start fires,
-            //   connectCount == 1, and there is NO network_restore_ride_through.
-            TMUX_CONNECT_ATTEMPTS.set(0)
-            val registry = ActiveTmuxClients()
-            val connector = QueueLeaseConnector(FakeSshSession())
-            val reconnectClient = FakeTmuxClient().withSinglePane("work", "%1")
-            val vm = newVm(
-                registry = registry,
-                sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-            )
-            vm.setTmuxClientFactoryForTest { _, _, _ -> reconnectClient }
-            vm.setAutoReconnectDelaysForTest(listOf(0L))
-            vm.replaceClientForTest(
-                hostId = 7L,
-                hostName = "alpha",
-                host = "alpha.example",
-                port = 22,
-                user = "alex",
-                keyPath = "/keys/a",
-                sessionName = "work",
-                client = FakeTmuxClient(),
-            )
-            runCurrent()
-
-            // The maintainer's exact cellular state: keepalive timestamp FRESH (the
-            // passive check would say "proven alive") but the post-handoff socket
-            // GENUINELY DEAD (the bounded probe cannot round-trip).
-            vm.forceTransportProvenAliveForTest = true
-            vm.forceLivenessProbeDeadForTest = true
-
-            val diagnostics = installRecordingDiagnosticSink()
-            try {
-                registry.lifecycleHooksSnapshot().single().onNetworkChanged(networkLoss())
-                runCurrent()
-                registry.lifecycleHooksSnapshot().single().onNetworkChanged(networkRestore())
-                advanceUntilIdle()
-
-                // GREEN load-bearing #1: the restore must NOT ride through onto the
-                // dead transport on a fresh-keepalive timestamp — that passive
-                // ride-through is the #1193 bug.
-                assertTrue(
-                    "a fresh-keepalive-but-DEAD-socket restore must NOT ride through the dead " +
-                        "transport on the passive timestamp (the #1193 cellular spurious drop); " +
-                        "events=${diagnostics.events.map { it.name }}",
-                    diagnostics.eventsNamed("network_restore_ride_through").isEmpty(),
-                )
-                // GREEN load-bearing #2: the bounded probe detected the dead socket and
-                // cleanly redialled via the #997 fresh lease.
-                assertEquals(
-                    "the dead-socket restore must redial via the bounded-probe fresh lease",
-                    1,
-                    connector.connectCount,
-                )
-                assertEquals(
-                    "the redial fires with the network-reconnect trigger",
-                    TmuxConnectTrigger.NetworkReconnect,
-                    vm.latestRestoreIntentSnapshot()?.trigger,
-                )
-                assertTrue(
-                    "the redial records the fast restore signal (not a vacuous pass); " +
-                        "events=${diagnostics.events.map { it.name }}",
-                    diagnostics.eventsNamed("network_restore_reconnect_start").isNotEmpty(),
-                )
-            } finally {
-                diagnostics.close()
-                vm.forceTransportProvenAliveForTest = null
-            }
-        }
-
-    @Test
-    fun issue1193NetworkRestoreFreshKeepaliveAndAliveSocketStillRidesThroughViaProbe() =
-        runTest(scheduler) {
-            // Issue #1193 NO-REGRESSION (G2 class coverage): the genuine happy cellular
-            // dip — a fresh keepalive timestamp AND a truly-alive socket. With the
-            // passive fast-path deleted the restore now proves liveness with the bounded
-            // ACTIVE probe; on a live socket the probe ANSWERS fast (well inside the 2s
-            // budget) so the session still RIDES THROUGH with no redial — the
-            // #974/#981/#1042 spurious-drop-suppression win is preserved. The
-            // distinguishing assertion is cause == "probe_answered" (NOT
-            // "transport_proven_alive"): even with a fresh keepalive the decision now
-            // runs through the probe, never the deleted passive gate.
-            TMUX_CONNECT_ATTEMPTS.set(0)
-            val registry = ActiveTmuxClients()
-            val connector = QueueLeaseConnector(FakeSshSession())
-            val reconnectClient = FakeTmuxClient().withSinglePane("work", "%1")
-            val vm = newVm(
-                registry = registry,
-                sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-            )
-            vm.setTmuxClientFactoryForTest { _, _, _ -> reconnectClient }
-            vm.setAutoReconnectDelaysForTest(listOf(0L))
-            vm.replaceClientForTest(
-                hostId = 7L,
-                hostName = "alpha",
-                host = "alpha.example",
-                port = 22,
-                user = "alex",
-                keyPath = "/keys/a",
-                sessionName = "work",
-                client = FakeTmuxClient(),
-            )
-            runCurrent()
-
-            // Fresh keepalive timestamp AND a live socket (probe-dead seam OFF, so the
-            // FakeTmuxClient's refresh-client round-trip answers).
-            vm.forceTransportProvenAliveForTest = true
-            vm.forceLivenessProbeDeadForTest = false
-
-            val diagnostics = installRecordingDiagnosticSink()
-            try {
-                registry.lifecycleHooksSnapshot().single().onNetworkChanged(networkLoss())
-                runCurrent()
-                registry.lifecycleHooksSnapshot().single().onNetworkChanged(networkRestore())
-                advanceUntilIdle()
-
-                assertEquals(
-                    "a fresh-keepalive + ALIVE-socket restore must still ride through (no redial)",
-                    0,
-                    connector.connectCount,
-                )
-                assertNull(
-                    "a ride-through restore must not schedule a network-reconnect redial",
-                    vm.latestRestoreIntentSnapshot(),
-                )
-                assertTrue(
-                    "the ride-through clears the loss-hold band back to Connected",
-                    vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected,
-                )
-                assertTrue(
-                    "no redial diagnostic may fire on the alive-socket ride-through; events=" +
-                        diagnostics.events.map { it.name },
-                    diagnostics.eventsNamed("network_restore_reconnect_start").isEmpty() &&
-                        diagnostics.eventsNamed("reconnect_start").isEmpty(),
-                )
-                val rideThrough = diagnostics.eventsNamed("network_restore_ride_through")
-                assertTrue(
-                    "expected a network_restore_ride_through (proves the ride-through fired, " +
-                        "not a vacuous pass); events=${diagnostics.events.map { it.name }}",
-                    rideThrough.isNotEmpty(),
-                )
-                assertEquals(
-                    "even with a fresh keepalive the restore now rides through via the bounded " +
-                        "PROBE (cause=probe_answered), NOT the deleted passive transport_proven_alive " +
-                        "fast-path — this is the #1193 behavior change",
-                    "probe_answered",
-                    rideThrough.single().fields["cause"],
-                )
-            } finally {
-                diagnostics.close()
-                vm.forceTransportProvenAliveForTest = null
-            }
-        }
-
-    // ---- Issue #1080: Doze/app-standby wake must NOT ride through a dead socket ----
-    //
-    // Android deep Doze suspends the network so the keepalive cannot keep the NAT
-    // mapping alive; the socket is silently dead on wake. The KEYSTONE fix is that
-    // the transport-liveness oracle (RealSshSession.isTransportProvenAliveWithin-
-    // KeepAliveWindow) now uses the wall-elapsed boot clock that COUNTS deep sleep,
-    // so after a real Doze gap it correctly reports STALE (proven at the core-ssh
-    // layer by RealSshSessionDozeClockTest). These VM tests pin the DOWNSTREAM
-    // consequence: when the oracle reports stale on a post-Doze restore the
-    // #1042/#1078 restore path must redial (issue ONE bounded probe, then a fresh
-    // lease), NOT ride through the dead socket; and when the oracle is still alive
-    // (a short background, not deep sleep) the restore must still ride through.
-
-    @Test
-    fun issue1080PostDozeWakeStaleTransportRedialsInsteadOfRidingThrough() = runTest(scheduler) {
-        // Post-deep-Doze wake: the NAT mapping died during sleep, and because the
-        // oracle now counts deep sleep (the #1080 boot-clock fix) it correctly
-        // reports STALE — model that with transportProvenAlive=false (oracle aged
-        // out) AND a DEAD bounded probe (the dead socket cannot answer). The restore
-        // path MUST fall through to a fresh-lease redial, NOT ride through. Without
-        // the #1080 clock fix the real on-device oracle would have falsely reported
-        // ALIVE here (frozen monotonic clock under-counting the sleep) and the
-        // restore would have ridden through the dead socket — the bug this guards.
-        TMUX_CONNECT_ATTEMPTS.set(0)
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val reconnectClient = FakeTmuxClient().withSinglePane("work", "%1")
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setTmuxClientFactoryForTest { _, sessionName, _ ->
-            assertEquals("work", sessionName)
-            reconnectClient
-        }
-        vm.setAutoReconnectDelaysForTest(listOf(0L))
-        // The live session whose keepalive aged out across the Doze gap (oracle stale).
-        val session = FakeSshSession().apply { transportProvenAlive = false }
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = FakeTmuxClient(),
-            session = session,
-        )
-        runCurrent()
-        // The dead post-Doze socket cannot answer the bounded restore probe either.
-        vm.forceLivenessProbeDeadForTest = true
-        assertFalse(
-            "precondition: a post-Doze stale transport must NOT report proven-alive",
-            vm.isTransportKeepAliveProvenAliveRecentlyForTest(),
-        )
-
-        // The Doze interval: network drops while backgrounded, then restores on wake.
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(networkLoss())
-        runCurrent()
-        assertTrue(
-            "the loss leaves the session in the loss-suspended Reconnecting state",
-            vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Reconnecting,
-        )
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(networkRestore())
-        advanceUntilIdle()
-
-        assertEquals(
-            "a post-Doze restore over a STALE/dead transport (oracle aged out + probe dead) " +
-                "MUST redial — it must NOT ride through the dead socket (#1080)",
-            TmuxConnectTrigger.NetworkReconnect,
-            vm.latestRestoreIntentSnapshot()?.trigger,
-        )
-        assertEquals("exactly one fresh-lease redial on the post-Doze restore", 1, connector.connectCount)
-        assertSame(reconnectClient, registry.clients.value[7L]?.client)
-
-        vm.forceLivenessProbeDeadForTest = false
-    }
-
-    @Test
-    fun issue1080ShortBackgroundWakeAliveTransportStillRidesThrough() = runTest(scheduler) {
-        // Class-coverage (G2): the #1080 fix must NOT turn every background wake into
-        // a reconnect. A short background (screen-off, app-switch — NOT deep sleep)
-        // leaves the keepalive proving the link alive; the oracle still reports ALIVE
-        // and the restore must RIDE THROUGH with zero redials, exactly as #1042.
-        TMUX_CONNECT_ATTEMPTS.set(0)
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val reconnectClient = FakeTmuxClient().withSinglePane("work", "%1")
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setTmuxClientFactoryForTest { _, _, _ -> reconnectClient }
-        vm.setAutoReconnectDelaysForTest(listOf(0L))
-        val session = FakeSshSession().apply { transportProvenAlive = true }
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = FakeTmuxClient(),
-            session = session,
-        )
-        runCurrent()
-        assertTrue(
-            "precondition: a short-background link is still proven-alive (no deep sleep)",
-            vm.isTransportKeepAliveProvenAliveRecentlyForTest(),
-        )
-
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(networkLoss())
-        runCurrent()
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(networkRestore())
-        advanceUntilIdle()
-
-        assertEquals(
-            "a wake over a PROVEN-ALIVE transport must ride through with no redial (#1080 must " +
-                "not over-reconnect a healthy short-background link)",
-            0,
-            connector.connectCount,
-        )
-        assertNull(
-            "a ride-through wake must not schedule a network-reconnect redial",
-            vm.latestRestoreIntentSnapshot(),
-        )
-        assertTrue(
-            "the ride-through clears the loss-hold band back to Connected",
-            vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected,
-        )
-    }
-
-    @Test
-    fun issue1078HandoffProbesAndRedialsWhenSocketDeadButPassivelyProvenAlive() = runTest(scheduler) {
-        // Issue #1078 reproduce-first (RED on base): the headline residual cellular
-        // stall. A REAL validated WIFI→CELLULAR handoff arrives while the transport is
-        // only PASSIVELY proven alive — the keepalive's last-inbound-byte age is still
-        // under the ~90s budget (forceTransportProvenAliveForTest=true → the reducer
-        // returns SuppressNetworkTransportProvenAlive) — but the old socket is GENUINELY
-        // DEAD post-handoff (an active probe fails, forceLivenessProbeDeadForTest=true).
-        //
-        // BASE BEHAVIOUR (the bug): the handoff suppress arm rode through on the passive
-        // timestamp ALONE with no active probe, so the session showed Live-but-FROZEN for
-        // up to ~90s until the keepalive budget finally tripped — no redial was ever
-        // scheduled. This test asserts a redial DOES fire, so it FAILS red on base.
-        //
-        // FIXED BEHAVIOUR (#1078, mirroring the #1042 restore arm): the suppress arm now
-        // issues ONE bounded active probe; it does not answer (dead socket), so it redials
-        // promptly via the SAME dead-link handoff authority (scheduleNetworkReconnect)
-        // within the probe budget instead of waiting out the ~90s keepalive.
-        TMUX_CONNECT_ATTEMPTS.set(0)
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val reconnectClient = FakeTmuxClient().withSinglePane("work", "%1")
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setTmuxClientFactoryForTest { _, sessionName, _ ->
-            assertEquals("work", sessionName)
-            reconnectClient
-        }
-        vm.setAutoReconnectDelaysForTest(listOf(0L))
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = FakeTmuxClient(),
-        )
-        runCurrent()
-        assertTrue(
-            "precondition: a healthy live session before the handoff",
-            vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected,
-        )
-
-        // Passively proven alive (recent inbound byte) but the active probe is DEAD:
-        // exactly the dead-socket-after-handoff state the keepalive timestamp masks.
-        vm.forceTransportProvenAliveForTest = true
-        vm.forceLivenessProbeDeadForTest = true
-
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(
-            networkChange(
-                previous = TerminalNetworkSnapshot.Validated("wifi"),
-                current = TerminalNetworkSnapshot.Validated("cell"),
-                previousValidated = TerminalNetworkSnapshot.Validated("wifi"),
-                reason = "wifi-cellular-handoff-dead-socket-recent-inbound",
-            ),
-        )
-        advanceUntilIdle()
-
-        assertEquals(
-            "a handoff whose bounded probe does NOT answer must redial promptly (not freeze " +
-                "Live for ~90s) — #1078",
-            TmuxConnectTrigger.NetworkReconnect,
-            vm.latestRestoreIntentSnapshot()?.trigger,
-        )
-        assertEquals(
-            "exactly one fresh-lease redial after the dead-socket handoff probe fails",
-            1,
-            connector.connectCount,
-        )
-        assertSame(reconnectClient, registry.clients.value[7L]?.client)
-
-        vm.forceTransportProvenAliveForTest = null
-        vm.forceLivenessProbeDeadForTest = false
-    }
-
-    @Test
-    fun issue1078HandoffRidesThroughWhenProbeAnswersOnAliveSocket() = runTest(scheduler) {
-        // Issue #1078 class-coverage (D32 G2): the alive-socket handoff case. Same
-        // WIFI→CELLULAR validated handoff while passively proven alive
-        // (forceTransportProvenAliveForTest=true), but here the bounded active probe
-        // ANSWERS (forceLivenessProbeDeadForTest=false → the FakeTmuxClient refresh-client
-        // round-trips). The genuine ride-through win (#981/#974/#1058) MUST be preserved:
-        // NO spurious redial, the session stays Connected, and the ride-through is
-        // attributed to the probe (probeConfirmed=true) so the pass is not vacuous.
-        TMUX_CONNECT_ATTEMPTS.set(0)
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val reconnectClient = FakeTmuxClient().withSinglePane("work", "%1")
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setTmuxClientFactoryForTest { _, _, _ -> reconnectClient }
-        vm.setAutoReconnectDelaysForTest(listOf(0L))
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = FakeTmuxClient(),
-        )
-        runCurrent()
-
-        vm.forceTransportProvenAliveForTest = true
-        vm.forceLivenessProbeDeadForTest = false
-
-        val diagnostics = installRecordingDiagnosticSink()
-        try {
-            registry.lifecycleHooksSnapshot().single().onNetworkChanged(
-                networkChange(
-                    previous = TerminalNetworkSnapshot.Validated("wifi"),
-                    current = TerminalNetworkSnapshot.Validated("cell"),
-                    previousValidated = TerminalNetworkSnapshot.Validated("wifi"),
-                    reason = "wifi-cellular-handoff-alive-socket",
-                ),
-            )
-            advanceUntilIdle()
-
-            assertEquals(
-                "a handoff whose bounded probe ANSWERS must NOT redial (rides through) — #1078",
-                0,
-                connector.connectCount,
-            )
-            assertNull(
-                "a probe-confirmed handoff ride-through must not schedule a network-reconnect redial",
-                vm.latestRestoreIntentSnapshot(),
-            )
-            assertTrue(
-                "the ride-through keeps the session Connected (no Reconnecting overlay)",
-                vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected,
-            )
-            assertTrue(
-                "no redial diagnostic may fire on the probe-confirmed handoff ride-through; events=" +
-                    diagnostics.events.map { it.name },
-                diagnostics.eventsNamed("network_reconnect_start").isEmpty() &&
-                    diagnostics.eventsNamed("reconnect_start").isEmpty(),
-            )
-            val rideThrough = diagnostics.eventsNamed("network_reconnect_skip").filter {
-                it.fields["cause"] == "transport_proven_alive"
-            }
-            assertTrue(
-                "expected a transport_proven_alive network_reconnect_skip (proves the ride-through " +
-                    "arm fired, not a vacuous pass); events=${diagnostics.events.map { it.name }}",
-                rideThrough.isNotEmpty(),
-            )
-            assertEquals(
-                "the ride-through must be attributed to the bounded ACTIVE probe answering, not " +
-                    "the passive keepalive timestamp alone (distinguishes #1078 from a passive suppress)",
-                true,
-                rideThrough.single().fields["probeConfirmed"],
-            )
-        } finally {
-            diagnostics.close()
-            vm.forceTransportProvenAliveForTest = null
-        }
-    }
-
-    @Test
-    fun networkReconnectAndPassiveDisconnectAreCoalescedIntoOneAttempt() = runTest(scheduler) {
-        TMUX_CONNECT_ATTEMPTS.set(0)
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val reconnectClient = FakeTmuxClient().withSinglePane("work", "%1")
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setTmuxClientFactoryForTest { _, sessionName, _ ->
-            assertEquals("work", sessionName)
-            reconnectClient
-        }
-        vm.setAutoReconnectDelaysForTest(listOf(60_000L))
-        val oldClient = FakeTmuxClient()
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = oldClient,
-        )
-        advanceUntilIdle()
-
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(
-            networkChange(reason = "wifi-cellular-handoff"),
-        )
-        runCurrent()
-        oldClient.disconnectedSignal.value = true
-        runCurrent()
-
-        assertTrue(
-            "passive EOF during a scheduled network reconnect must not replace it with a second loop",
-            vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Reconnecting,
-        )
-        assertEquals(0, connector.connectCount)
-        assertEquals(0, TMUX_CONNECT_ATTEMPTS.get())
-
-        advanceTimeBy(60_000L)
-        advanceUntilIdle()
-
-        assertEquals(
-            "only the already scheduled network reconnect should open a transport",
-            1,
-            connector.connectCount,
-        )
-        assertEquals(1, TMUX_CONNECT_ATTEMPTS.get())
-        assertSame(reconnectClient, registry.clients.value[7L]?.client)
-    }
-
-    @Test
-    fun networkChangeLifecycleHookProactivelyReattachesTmuxSession() = runTest(scheduler) {
-        val registry = ActiveTmuxClients()
-        val connector = QueueLeaseConnector(FakeSshSession())
-        val reconnectClient = FakeTmuxClient().withSinglePane("work", "%1")
-        val vm = newVm(
-            registry = registry,
-            sshLeaseManager = testLeaseManager(connector = connector, scope = this, idleTtlMillis = 0L),
-        )
-        vm.setTmuxClientFactoryForTest { _, sessionName, _ ->
-            assertEquals("work", sessionName)
-            reconnectClient
-        }
-        vm.setAutoReconnectDelaysForTest(listOf(0L))
-        val oldClient = FakeTmuxClient()
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = oldClient,
-        )
-
-        registry.lifecycleHooksSnapshot().single().onNetworkChanged(
-            networkChange(reason = "validated-default-network-changed"),
-        )
-        advanceUntilIdle()
-
-        assertEquals(1, connector.connectCount)
-        assertTrue("old tmux control client must be detached during reconnect", oldClient.detachCleanlyCalled)
-        assertSame(reconnectClient, registry.clients.value[7L]?.client)
-        assertTrue(vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected)
-        assertEquals(TmuxConnectTrigger.NetworkReconnect, vm.latestRestoreIntentSnapshot()?.trigger)
-    }
-
-    @Test
     fun networkReconnectEvictsConnectedIdleLeaseBeforeReattaching() = runTest(scheduler) {
         val registry = ActiveTmuxClients()
         val staleSession = FakeSshSession()
@@ -6359,483 +4956,6 @@ class TmuxSessionViewModelTest : TmuxSessionViewModelTestBase() {
         confidence = AgentDetection.Confidence.ProcessConfirmed,
     )
 
-    // ─── Issue #793: tail-first conversation load + load state machine ───
-
-    @Test
-    fun openingLargeConversationLoadsTailWithoutFetchingWholeHistory() = runTest(scheduler) {
-        // REGRESSION PROOF (#793): a LARGE transcript (5000 raw lines on the
-        // server). Opening the Conversation tab must render the TAIL quickly
-        // WITHOUT fetching the whole history first, and report that older
-        // messages remain to page in.
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        val detection = newClaudeDetection()
-        // The tail window the server returns: only the most recent turns.
-        val tailJsonl = listOf(
-            """{"type":"user","uuid":"u4999","message":{"role":"user","content":"the latest question"}}""",
-            """{"type":"assistant","uuid":"a4999","message":{"role":"assistant","content":"the latest answer"}}""",
-        ).joinToString("\n")
-        val session = FakeSshSession(
-            wcOutput = "5000\n",
-            initialEventsOutput = tailJsonl,
-        )
-
-        vm.startAgentConversationForPaneForTest(
-            paneId = "%0",
-            session = session,
-            detection = detection,
-        )
-        advanceUntilIdle()
-
-        val state = vm.agentConversations.value["%0"]!!
-        // The tail rendered.
-        assertEquals(
-            listOf("the latest question", "the latest answer"),
-            state.events.filterIsInstance<ConversationEvent.Message>().map { it.text },
-        )
-        // Load resolved to a clear terminal state (Ready), NOT a stuck spinner.
-        assertEquals(ConversationLoadState.Ready, state.loadState)
-        // Older messages remain → the pane offers upward paging.
-        assertTrue("older messages must be pageable", state.hasMoreOlderEvents)
-
-        // The CORE assertion: the open did NOT fetch the whole history. The
-        // only window read is capped at the first-paint raw budget
-        // (FIRST_PAINT_MESSAGE_BUDGET * JSONL_RAW_LINES_PER_EVENT raw lines),
-        // far below the 5000-line file.
-        val windowCommand = session.execCommands.single { it.contains("@@PS_WINDOW@@") }
-        val firstPaintRawBudget = FIRST_PAINT_MESSAGE_BUDGET * JSONL_RAW_LINES_PER_EVENT
-        assertTrue(
-            "expected a first-paint-budget tail; got $windowCommand",
-            windowCommand.contains("tail -n $firstPaintRawBudget "),
-        )
-        // No wide 500-message (=4000 raw line) eager read was issued.
-        val eagerRawBudget = 500 * JSONL_RAW_LINES_PER_EVENT
-        assertFalse(
-            "the eager full-history read path must be gone (D22 hard-cut)",
-            session.execCommands.any { it.contains("tail -n $eagerRawBudget ") },
-        )
-    }
-
-    @Test
-    fun coldOpenDropsRedundantLineCountExecAndTailsFromWindowPosition() = runTest(scheduler) {
-        // REGRESSION PROOF (#817 slice 1): the cold-open path must NOT issue a
-        // separate `wc -l` (lineCount) round-trip before the windowed read. The
-        // window read already yields the file's line count, and the follow-tail
-        // starts from THAT position (no dropped/duplicated events).
-        TmuxSessionLatencyTelemetry.resetForTest()
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        val detection = newClaudeDetection()
-        val tailJsonl = listOf(
-            """{"type":"user","uuid":"u1","message":{"role":"user","content":"latest question"}}""",
-            """{"type":"assistant","uuid":"a1","message":{"role":"assistant","content":"latest answer"}}""",
-        ).joinToString("\n")
-        val session = FakeSshSession(
-            wcOutput = "1234\n",
-            initialEventsOutput = tailJsonl,
-        )
-
-        vm.startAgentConversationForPaneForTest(
-            paneId = "%0",
-            session = session,
-            detection = detection,
-        )
-        advanceUntilIdle()
-
-        // Events loaded correctly from the window read.
-        val state = vm.agentConversations.value["%0"]!!
-        assertEquals(
-            listOf("latest question", "latest answer"),
-            state.events.filterIsInstance<ConversationEvent.Message>().map { it.text },
-        )
-
-        // (a) The redundant standalone `wc -l` (lineCount) exec is gone. The
-        // ONLY count of the source file happens inside the windowed read's own
-        // single exec (the @@PS_WINDOW@@ sentinel command), never as a separate
-        // `wc -l < ...` round-trip.
-        val standaloneWcExecs = session.execCommands.filter {
-            it.contains("wc -l < ") && !it.contains("@@PS_WINDOW@@")
-        }
-        assertTrue(
-            "cold open must not fire a separate lineCount `wc -l` exec; commands=${session.execCommands}",
-            standaloneWcExecs.isEmpty(),
-        )
-
-        // (b) The follow-tail starts from the window read's line count (1234),
-        // so it only emits lines appended AFTER the window — no gaps/dupes.
-        assertEquals(
-            listOf("/home/u/.claude/sessions/abc.jsonl" to 1234L),
-            session.tailFromLineCalls,
-        )
-
-        // The conversation_open latency span was emitted with the open duration
-        // so a connected test / logcat can snapshot the authoritative number.
-        val openEvents = TmuxSessionLatencyTelemetry.snapshot()
-            .filter { it.name == CONVERSATION_OPEN_LATENCY_OPERATION }
-        assertEquals(
-            "exactly one conversation_open span; spans=${TmuxSessionLatencyTelemetry.snapshot().map { it.name }}",
-            1,
-            openEvents.size,
-        )
-        assertEquals("%0", openEvents.single().paneId)
-        assertTrue(
-            "span must carry a grep-able artifact line; got ${openEvents.single().toArtifactLine()}",
-            openEvents.single().toArtifactLine().contains("tmux_latency_conversation_open_ms="),
-        )
-        TmuxSessionLatencyTelemetry.resetForTest()
-    }
-
-    @Test
-    fun openingExistingConversationStartsInLoadingThenReady() = runTest(scheduler) {
-        // Problem 1/2 (#793): the Conversation tab shows a "Loading
-        // conversation…" state (loadState == Loading) while the tail read is in
-        // flight, and resolves to Ready — never a stuck "Waiting for agent…".
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        vm.applyParsedPanesForTest(
-            listOf(TmuxSessionViewModel.ParsedPane("%0", "@0", "$0", "node", paneIndex = 0)),
-        )
-        val detection = newClaudeDetection()
-        val execGate = CompletableDeferred<Unit>()
-        val session = FakeSshSession(
-            wcOutput = "10\n",
-            initialEventsOutput =
-                """{"type":"assistant","uuid":"a1","message":{"role":"assistant","content":"hi"}}""",
-            execGate = execGate,
-        )
-
-        // Realistic flow: the user opens the Conversation tab (seeds a Loading
-        // placeholder row), THEN detection lands and the transcript loads.
-        vm.selectSessionTab("%0", SessionTab.Conversation)
-        runCurrent()
-        assertEquals(ConversationLoadState.Loading, vm.agentConversations.value["%0"]!!.loadState)
-
-        val loadJob = backgroundScope.launch {
-            vm.startAgentConversationForPaneForTest("%0", session, detection)
-        }
-        runCurrent()
-        // While the read is gated (in flight), the row is still Loading.
-        assertEquals(ConversationLoadState.Loading, vm.agentConversations.value["%0"]!!.loadState)
-
-        // Release the read; it resolves to Ready.
-        execGate.complete(Unit)
-        advanceUntilIdle()
-        loadJob.join()
-        assertEquals(ConversationLoadState.Ready, vm.agentConversations.value["%0"]!!.loadState)
-    }
-
-    @Test
-    fun conversationLoadWatchdogFlipsLoadingToFailedSoItNeverSpinsForever() = runTest(scheduler) {
-        // Problem 2 (#793): a never-completing load (the transport flap symptom)
-        // must surface a clear Failed terminal state, not an infinite spinner.
-        // Drive the detection-pending placeholder path (selectSessionTab seeds a
-        // detection-less Loading row + arms the watchdog).
-        val vm = newVm()
-        vm.setConversationLoadTimeoutForTest(1_000L)
-        vm.attachClientForTest(FakeTmuxClient())
-        vm.applyParsedPanesForTest(
-            listOf(TmuxSessionViewModel.ParsedPane("%0", "@0", "$0", "node", paneIndex = 0)),
-        )
-
-        vm.selectSessionTab("%0", SessionTab.Conversation)
-        runCurrent()
-        assertEquals(ConversationLoadState.Loading, vm.agentConversations.value["%0"]!!.loadState)
-
-        // Time passes with no detection/transcript ever landing.
-        advanceTimeBy(1_500L)
-        runCurrent()
-        assertEquals(
-            "the watchdog must flip a never-completing load to Failed",
-            ConversationLoadState.Failed,
-            vm.agentConversations.value["%0"]!!.loadState,
-        )
-    }
-
-    @Test
-    fun emptyTranscriptResolvesToEmptyTerminalStateNotSpinner() = runTest(scheduler) {
-        // Problem 2 (#793): a successful read of an empty transcript surfaces a
-        // clear Empty state, not a stuck spinner.
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        val detection = newClaudeDetection()
-        val session = FakeSshSession(wcOutput = "0\n", initialEventsOutput = "")
-
-        vm.startAgentConversationForPaneForTest("%0", session, detection)
-        advanceUntilIdle()
-
-        val state = vm.agentConversations.value["%0"]!!
-        assertEquals(ConversationLoadState.Empty, state.loadState)
-        assertTrue(state.events.isEmpty())
-        assertFalse(state.hasMoreOlderEvents)
-    }
-
-    @Test
-    fun loadingOlderEventsWidensWindowAndPrependsOlderMessages() = runTest(scheduler) {
-        // AC3 (#793): older messages page in lazily on upward scroll, preserving
-        // the already-loaded tail (the merge keeps newer events in place).
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        val detection = newClaudeDetection()
-        // First paint: a large file, small tail.
-        val firstPaintTail =
-            """{"type":"user","uuid":"u2","message":{"role":"user","content":"recent"}}"""
-        val session = FakeSshSession(
-            wcOutput = "5000\n",
-            initialEventsOutput = firstPaintTail,
-        )
-        // loadOlderAgentConversationEvents reads sessionRef + activeTarget, so
-        // wire the same session as the live runtime session for the pane.
-        vm.attachSessionForAgentRetryForTest(session)
-        vm.startAgentConversationForPaneForTest("%0", session, detection)
-        advanceUntilIdle()
-        assertTrue(vm.agentConversations.value["%0"]!!.hasMoreOlderEvents)
-
-        // The page-older read returns a WIDER window that includes an older
-        // message AHEAD of the recent one.
-        session.setInitialEventsOutput(
-            listOf(
-                """{"type":"user","uuid":"u1","message":{"role":"user","content":"older"}}""",
-                """{"type":"user","uuid":"u2","message":{"role":"user","content":"recent"}}""",
-            ).joinToString("\n"),
-        )
-        session.setWcOutput("12\n") // now the wider window covers the whole file
-
-        vm.loadOlderAgentConversationEvents("%0")
-        advanceUntilIdle()
-
-        val state = vm.agentConversations.value["%0"]!!
-        assertEquals(
-            "older message must be prepended ahead of the recent one",
-            listOf("older", "recent"),
-            state.events.filterIsInstance<ConversationEvent.Message>().map { it.text },
-        )
-        assertFalse("the whole file is now in the window", state.hasMoreOlderEvents)
-        assertFalse(state.isPagingOlder)
-    }
-
-    @Test
-    fun detectedAgentConversationStartsWithoutHintPopupState() = runTest(scheduler) {
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        vm.applyParsedPanesForTest(
-            listOf(TmuxSessionViewModel.ParsedPane("%0", "@0", "$0", "shell", paneIndex = 0)),
-        )
-
-        vm.startAgentConversationForTest("%0", newClaudeDetection())
-
-        val state = vm.agentConversations.value["%0"]
-        assertNotNull(state)
-        assertEquals(SessionTab.Terminal, state!!.selectedTab)
-        assertEquals(AgentKind.ClaudeCode, state.detection?.agent)
-    }
-
-    @Test
-    fun agentConversationStillReceivesEventsAfterPopupRemoval() = runTest(scheduler) {
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        vm.startAgentConversationForTest("%0", newClaudeDetection())
-
-        vm.appendAgentEventsForTest(
-            "%0",
-            listOf(
-                ConversationEvent.Message(
-                    id = "assistant-1",
-                    agent = AgentKind.ClaudeCode,
-                    atMillis = 1L,
-                    role = ConversationRole.Assistant,
-                    text = "still here",
-                ),
-            ),
-        )
-
-        val after = vm.agentConversations.value["%0"]!!
-        assertEquals(
-            "the new event should reach the conversation feed",
-            "assistant-1",
-            after.events.last().id,
-        )
-        assertEquals(AgentKind.ClaudeCode, after.detection?.agent)
-    }
-
-    // ─── Issue #256: current-pane conversation semantics ──────────────
-
-    @Test
-    fun selectingConversationTabOnlyMutatesTheCurrentPaneState() = runTest(scheduler) {
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        vm.startAgentConversationForTest("%0", newClaudeDetection())
-        vm.startAgentConversationForTest(
-            "%1",
-            AgentDetection(
-                agent = AgentKind.Codex,
-                sourcePath = "/home/u/.codex/sessions/xyz.jsonl",
-                sessionId = "xyz",
-                confidence = AgentDetection.Confidence.ProcessConfirmed,
-            ),
-        )
-
-        vm.selectSessionTab("%0", SessionTab.Conversation)
-
-        assertEquals(SessionTab.Conversation, vm.agentConversations.value["%0"]!!.selectedTab)
-        assertEquals(SessionTab.Terminal, vm.agentConversations.value["%1"]!!.selectedTab)
-    }
-
-    @Test
-    fun warmSwitchToLoadedConversationRecordsConversationSwitchSpan() = runTest(scheduler) {
-        // Issue #817 (Rank-1 measurement): switching Terminal -> Conversation on a
-        // row whose transcript is ALREADY loaded (events present) records a
-        // conversation_switch latency span. This is the warm-switch case the spike
-        // predicted is already <0.3s (a pure StateFlow read, no SSH); the span
-        // makes that an authoritative number.
-        TmuxSessionLatencyTelemetry.resetForTest()
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        val event = ConversationEvent.Message(
-            id = "message-1",
-            agent = AgentKind.ClaudeCode,
-            role = ConversationRole.Assistant,
-            text = "already loaded answer",
-        )
-        // Seed a row with events, default tab Terminal (the #815 default).
-        vm.startAgentConversationForTest("%0", newClaudeDetection(), listOf(event))
-        assertEquals(SessionTab.Terminal, vm.agentConversations.value["%0"]!!.selectedTab)
-
-        vm.selectSessionTab("%0", SessionTab.Conversation)
-
-        val switchEvents = TmuxSessionLatencyTelemetry.snapshot()
-            .filter { it.name == CONVERSATION_SWITCH_LATENCY_OPERATION }
-        assertEquals(
-            "exactly one conversation_switch span; spans=" +
-                TmuxSessionLatencyTelemetry.snapshot().map { it.name },
-            1,
-            switchEvents.size,
-        )
-        assertEquals("%0", switchEvents.single().paneId)
-        assertTrue(
-            "span must carry a grep-able artifact line; got ${switchEvents.single().toArtifactLine()}",
-            switchEvents.single().toArtifactLine().contains("tmux_latency_conversation_switch_ms="),
-        )
-        TmuxSessionLatencyTelemetry.resetForTest()
-    }
-
-    @Test
-    fun switchingToConversationOnAStillLoadingRowDoesNotRecordSwitchSpan() = runTest(scheduler) {
-        // Issue #817: a tap that lands on a still-loading/placeholder row (no
-        // events yet) is the OPEN path, NOT the warm switch — it must NOT emit a
-        // conversation_switch span (the open is covered by conversation_open_full).
-        TmuxSessionLatencyTelemetry.resetForTest()
-        val vm = newVm()
-        vm.connectWithPaneForTest(paneId = "%0", windowId = "@0")
-        runCurrent()
-
-        // selectSessionTab seeds a detection-less, event-less Loading placeholder.
-        vm.selectSessionTab("%0", SessionTab.Conversation)
-
-        val switchEvents = TmuxSessionLatencyTelemetry.snapshot()
-            .filter { it.name == CONVERSATION_SWITCH_LATENCY_OPERATION }
-        assertTrue(
-            "a switch onto an empty/loading row must not record a warm-switch span; " +
-                "spans=${TmuxSessionLatencyTelemetry.snapshot().map { it.name }}",
-            switchEvents.isEmpty(),
-        )
-        TmuxSessionLatencyTelemetry.resetForTest()
-    }
-
-    @Test
-    fun selectingConversationTabOnPresumedAgentWithoutDetectionSeedsPlaceholderRow() = runTest(scheduler) {
-        // Issue #778: tapping Conversation on a live presumed-agent pane that has
-        // no conversation row yet (no live detection, no remembered status) must
-        // NOT be a no-op. It seeds a detection-less placeholder row with
-        // `selectedTab = Conversation` so the screen switches to the Conversation
-        // surface (placeholder) instead of staying stuck on Terminal. The real
-        // transcript replaces this when detection lands.
-        val vm = newVm()
-        // Issue #878: pin the open-time default to Terminal so the pane-add
-        // auto-seed no-ops and this test isolates the #778 TAP-to-seed path
-        // (with the Conversation default the row would already exist from the
-        // #878 auto-seed, which a separate test covers).
-        vm.setDefaultAgentSessionViewForTest(
-            com.pocketshell.app.settings.DefaultAgentSessionView.Terminal,
-        )
-        vm.connectWithPaneForTest(paneId = "%0", windowId = "@0")
-        runCurrent()
-        assertNull(vm.agentConversations.value["%0"])
-
-        vm.selectSessionTab("%0", SessionTab.Conversation)
-
-        val row = vm.agentConversations.value["%0"]
-        assertNotNull(row)
-        assertEquals(SessionTab.Conversation, row!!.selectedTab)
-        assertNull(row.detection)
-    }
-
-    @Test
-    fun selectingConversationTabOnExistingDetectionlessRowSwitchesToConversation() = runTest(scheduler) {
-        // Issue #778: a presumed-agent pane that already has a detection-less row
-        // (seeded by an earlier tap) honours a Conversation tap — it switches to
-        // Conversation even though `detection == null`, rather than swallowing it.
-        val vm = newVm()
-        vm.connectWithPaneForTest(paneId = "%0", windowId = "@0")
-        runCurrent()
-        // Seed a detection-less Conversation row via the seed-on-tap path, then
-        // flip back to Terminal so we can re-assert the Conversation switch.
-        vm.selectSessionTab("%0", SessionTab.Conversation)
-        vm.selectSessionTab("%0", SessionTab.Terminal)
-        assertEquals(SessionTab.Terminal, vm.agentConversations.value["%0"]!!.selectedTab)
-        assertNull(vm.agentConversations.value["%0"]!!.detection)
-
-        vm.selectSessionTab("%0", SessionTab.Conversation)
-
-        assertEquals(SessionTab.Conversation, vm.agentConversations.value["%0"]!!.selectedTab)
-        assertNull(vm.agentConversations.value["%0"]!!.detection)
-    }
-
-    @Test
-    fun selectingTmuxConversationAndTerminalTabsRecordsExplicitDiagnostics() = runTest(scheduler) {
-        val diagnostics = installRecordingDiagnosticSink()
-        try {
-            val vm = newVm()
-            vm.attachClientForTest(FakeTmuxClient())
-            val event = ConversationEvent.Message(
-                id = "message-1",
-                agent = AgentKind.ClaudeCode,
-                role = ConversationRole.Assistant,
-                text = "do not record this transcript text",
-            )
-            vm.startAgentConversationForTest("%0", newClaudeDetection(), listOf(event))
-
-            vm.selectSessionTab("%0", SessionTab.Conversation)
-            vm.selectSessionTab("%0", SessionTab.Terminal)
-
-            val events = diagnostics.eventsNamed("conversation_terminal_tab_switch")
-            assertEquals(2, events.size)
-            assertEquals("tmux", events[0].fields["mode"])
-            assertEquals("%0", events[0].fields["paneId"])
-            assertEquals("Terminal", events[0].fields["fromTab"])
-            assertEquals("Conversation", events[0].fields["toTab"])
-            assertEquals("terminal_to_conversation", events[0].fields["direction"])
-            assertEquals(true, events[0].fields["hasConversation"])
-            assertEquals(1, events[0].fields["eventCount"])
-            assertEquals("Conversation", events[1].fields["fromTab"])
-            assertEquals("Terminal", events[1].fields["toTab"])
-            assertFalse(events[0].fields.containsValue(event.text))
-        } finally {
-            diagnostics.close()
-        }
-    }
-
-    @Test
-    fun selectingConversationTabForUnknownOrPlainPaneIsNoOp() = runTest(scheduler) {
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        vm.startAgentConversationForTest("%0", newClaudeDetection())
-        val before = vm.agentConversations.value
-
-        vm.selectSessionTab("%missing", SessionTab.Conversation)
-
-        assertEquals(before, vm.agentConversations.value)
-    }
-
     // ─── Issue #495: remember agent sessions across reconnect ─────────
 
     /**
@@ -6869,143 +4989,6 @@ class TmuxSessionViewModelTest : TmuxSessionViewModelTestBase() {
                     sessionName = sessionName,
                 ),
             ),
-        )
-    }
-
-    @Test
-    fun agentSessionRememberedAndRestoredImmediatelyAfterReconnect() = runTest(scheduler) {
-        val vm = newVm()
-        vm.connectWithPaneForTest(paneId = "%0", windowId = "@0")
-        vm.startAgentConversationForTest("%0", newClaudeDetection())
-        // The user opens Conversation — this remembers the window's status.
-        vm.selectSessionTab("%0", SessionTab.Conversation)
-        runCurrent()
-
-        // Reconnect: tmux re-attach assigns a NEW pane id under the SAME
-        // window. The old pane row is gone; nothing is seeded by the live
-        // detection round-trip yet.
-        vm.applyParsedPanesForTest(
-            listOf(TmuxSessionViewModel.ParsedPane("%7", "@0", "$0", "shell", paneIndex = 0, sessionName = "work")),
-        )
-        runCurrent()
-
-        // Issue #819 (A2): the Conversation tab is available immediately on the
-        // new pane id (a row exists on the remembered tab), but the seed now
-        // restores a RESOLVING placeholder — NOT the remembered (possibly
-        // stale/sibling) detection rendered Live. The screen surfaces this via
-        // the `presumedAgent` "Loading conversation…" placeholder path (a
-        // detection-less row whose pane is not a confirmed shell), so the #495
-        // tab-availability benefit is preserved without flashing a stale source.
-        assertNull("old pane id is gone after reattach", vm.agentConversations.value["%0"])
-        val restored = vm.agentConversations.value["%7"]
-        assertNotNull("agent status restored immediately on the new pane", restored)
-        assertEquals(
-            "the remembered window is restored on the Conversation tab the user was on",
-            SessionTab.Conversation,
-            restored!!.selectedTab,
-        )
-        assertNull(
-            "#819 (A2): the seed must NOT carry the remembered (possibly stale) " +
-                "source — live detection re-anchors the route-true source",
-            restored.detection,
-        )
-        assertTrue(
-            "#819 (A2): the reattach seed is a remembered-agent resolving placeholder",
-            restored.rememberedAgentPlaceholder,
-        )
-        assertEquals(
-            "#819 (A2): the resolving placeholder holds Loading until live detection binds",
-            ConversationLoadState.Loading,
-            restored.loadState,
-        )
-    }
-
-    @Test
-    fun sharedAgentSessionMemoryRestoresAfterViewModelRecreation() = runTest(scheduler) {
-        val memory = AgentSessionMemory()
-        val firstVm = newVm(agentSessionMemory = memory)
-        firstVm.connectWithPaneForTest(paneId = "%0", windowId = "@0")
-        firstVm.startAgentConversationForTest("%0", newClaudeDetection())
-        firstVm.selectSessionTab("%0", SessionTab.Conversation)
-        runCurrent()
-
-        val recreatedVm = newVm(agentSessionMemory = memory)
-        recreatedVm.connectWithPaneForTest(paneId = "%8", windowId = "@0")
-        runCurrent()
-
-        val restored = recreatedVm.agentConversations.value["%8"]
-        assertNotNull(
-            "process-scoped memory must restore agent status for a fresh VM before detection",
-            restored,
-        )
-        assertEquals(
-            "the user's Conversation tab choice survives VM recreation",
-            SessionTab.Conversation,
-            restored!!.selectedTab,
-        )
-        // Issue #819 (A2): a cross-VM restore must also NOT trust the remembered
-        // source blind — it seeds the resolving placeholder and re-anchors the
-        // route-true source on live re-detection in the fresh VM.
-        assertNull(
-            "#819 (A2): the cross-VM seed must not carry the remembered source",
-            restored.detection,
-        )
-        assertTrue(
-            "#819 (A2): the cross-VM reattach seed is a remembered-agent resolving placeholder",
-            restored.rememberedAgentPlaceholder,
-        )
-    }
-
-    @Test
-    fun reconnectReselectsConversationWhenUserWasOnConversation() = runTest(scheduler) {
-        val vm = newVm()
-        vm.connectWithPaneForTest(paneId = "%0", windowId = "@0")
-        vm.startAgentConversationForTest("%0", newClaudeDetection())
-        vm.selectSessionTab("%0", SessionTab.Conversation)
-        runCurrent()
-
-        vm.applyParsedPanesForTest(
-            listOf(TmuxSessionViewModel.ParsedPane("%7", "@0", "$0", "shell", paneIndex = 0, sessionName = "work")),
-        )
-        runCurrent()
-
-        assertEquals(
-            "user who was in Conversation stays on Conversation after reconnect",
-            SessionTab.Conversation,
-            vm.agentConversations.value["%7"]!!.selectedTab,
-        )
-    }
-
-    @Test
-    fun reconnectKeepsTerminalWhenUserWasOnTerminal() = runTest(scheduler) {
-        val vm = newVm()
-        vm.connectWithPaneForTest(paneId = "%0", windowId = "@0")
-        vm.startAgentConversationForTest("%0", newClaudeDetection())
-        // User saw the agent but stayed on Terminal — remember Terminal.
-        vm.selectSessionTab("%0", SessionTab.Terminal)
-        runCurrent()
-
-        vm.applyParsedPanesForTest(
-            listOf(TmuxSessionViewModel.ParsedPane("%7", "@0", "$0", "shell", paneIndex = 0, sessionName = "work")),
-        )
-        runCurrent()
-
-        val restored = vm.agentConversations.value["%7"]!!
-        // Issue #819 (A2): the seed restores the remembered TAB (Terminal here)
-        // as a resolving placeholder — not the remembered (possibly stale)
-        // source — and live detection re-anchors the route-true source.
-        assertNull(
-            "#819 (A2): the seed must not carry the remembered source",
-            restored.detection,
-        )
-        assertTrue(
-            "#819 (A2): the reattach seed is a remembered-agent resolving placeholder",
-            restored.rememberedAgentPlaceholder,
-        )
-        assertEquals(
-            "Conversation tab available but Terminal stays selected",
-            SessionTab.Terminal,
-            restored.selectedTab,
         )
     }
 
@@ -7043,151 +5026,6 @@ class TmuxSessionViewModelTest : TmuxSessionViewModelTestBase() {
             sessionId = sessionId,
             confidence = AgentDetection.Confidence.ProcessConfirmed,
         )
-
-    @Test
-    fun codexReattachSeedDoesNotRenderRememberedSiblingSourceAsLive() = runTest(scheduler) {
-        // The reported scenario: a Codex window the user was reading in
-        // Conversation. The remembered detection was captured during a prior
-        // same-cwd mis-pick, so it carries the SIBLING/orchestrator rollout
-        // (`git-pocketshell-desktop-codex`), NOT the route's own session
-        // (`ai-shipping-labs`).
-        val siblingRollout = "/home/u/.codex/sessions/git-pocketshell-desktop-codex.jsonl"
-        val routeTrueRollout = "/home/u/.codex/sessions/ai-shipping-labs.jsonl"
-
-        val vm = newVm()
-        vm.connectWithPaneForTest(paneId = "%0", windowId = "@0")
-        // The remembered detection points at the SIBLING rollout (the stale
-        // mis-pick). Recording it + opening Conversation arms agentSessionMemory.
-        vm.startAgentConversationForTest(
-            "%0",
-            codexDetection(siblingRollout, "git-pocketshell-desktop-codex"),
-        )
-        vm.selectSessionTab("%0", SessionTab.Conversation)
-        runCurrent()
-
-        // Reconnect: a NEW pane id under the SAME window @0; the seed fires.
-        vm.applyParsedPanesForTest(
-            listOf(
-                TmuxSessionViewModel.ParsedPane(
-                    "%7", "@0", "$0", "shell", paneIndex = 0,
-                    cwd = "/home/u/git/pocketshell", paneTty = "/dev/pts/9",
-                    sessionName = "work",
-                ),
-            ),
-        )
-        runCurrent()
-
-        val seeded = vm.agentConversations.value["%7"]!!
-        // The LOAD-BEARING assertion (G6): the seed must NOT render the stale
-        // sibling source. On base the seed carries remembered.detection, so
-        // `seeded.detection!!.sourcePath == siblingRollout` (RED).
-        assertNull(
-            "#819 (A2): the reattach seed must NOT render the remembered " +
-                "(stale sibling) Codex source — it holds resolving until live " +
-                "re-detection binds the route-true source",
-            seeded.detection,
-        )
-        assertNotEquals(
-            "#819 (A2): the stale sibling rollout must NOT be the active source",
-            siblingRollout,
-            seeded.detection?.sourcePath,
-        )
-        assertTrue(seeded.rememberedAgentPlaceholder)
-        assertEquals(SessionTab.Conversation, seeded.selectedTab)
-
-        // Live re-detection (the route's OWN fd-pinned source via Slice A1) lands
-        // and binds the route-true rollout. The Conversation now matches the
-        // route-named Terminal.
-        vm.markAgentTailLiveForTest("%7", codexDetection(routeTrueRollout, "ai-shipping-labs"))
-        runCurrent()
-
-        val bound = vm.agentConversations.value["%7"]!!
-        assertEquals(
-            "#819 (A2): after live re-detection the Conversation is bound to the " +
-                "ROUTE-TRUE source the Terminal is attached to",
-            routeTrueRollout,
-            bound.detection!!.sourcePath,
-        )
-        assertEquals("ai-shipping-labs", bound.detection!!.sessionId)
-        assertEquals(
-            "the remembered Conversation tab is preserved across the re-anchor",
-            SessionTab.Conversation,
-            bound.selectedTab,
-        )
-        assertFalse(
-            "the resolving-placeholder flag clears once the route-true source binds",
-            bound.rememberedAgentPlaceholder,
-        )
-    }
-
-    @Test
-    fun cachedRecordedKindWithNullSourceRereadsSourceOptionAndRebinds() = runTest(scheduler) {
-        val now = System.currentTimeMillis() / 1000
-        val ownSource = "/home/u/.claude/projects/-workspace-proj/own.jsonl"
-        val siblingSource = "/home/u/.claude/projects/-workspace-proj/sibling.jsonl"
-        val session = FakeSshSession(
-            recordedKindOutput = "claude\n",
-            recordedSourceOutput = "",
-            detectionOutput = """
-                claude|${now - 60}|/workspace/proj|$ownSource
-                claude|$now|/workspace/proj|$siblingSource
-            """.trimIndent(),
-        )
-        val vm = newVm()
-        vm.replaceClientForTest(
-            hostId = 1L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = FakeTmuxClient(),
-            session = session,
-        )
-        vm.applyParsedPanesForTest(
-            listOf(
-                TmuxSessionViewModel.ParsedPane(
-                    paneId = "%0",
-                    windowId = "@0",
-                    sessionId = "$0",
-                    title = "claude",
-                    paneIndex = 0,
-                    cwd = "/workspace/proj",
-                    currentCommand = "claude",
-                    paneTty = "/dev/pts/7",
-                    sessionName = "work",
-                ),
-            ),
-        )
-        advanceUntilIdle()
-
-        assertEquals(
-            "precondition: before the delayed source option exists, the recorded-kind " +
-                "cache has source=null and Claude falls back to newest same-cwd candidate",
-            siblingSource,
-            vm.agentConversations.value["%0"]?.detection?.sourcePath,
-        )
-
-        session.setRecordedSourceOutput("$ownSource\n")
-        vm.startAgentDetectionForPaneForTest("%0")
-        advanceUntilIdle()
-
-        assertEquals(
-            "a cached recorded kind with source=null must re-read @ps_agent_source " +
-                "and rebind to the exact recorded source once the watcher writes it",
-            ownSource,
-            vm.agentConversations.value["%0"]?.detection?.sourcePath,
-        )
-        assertTrue(
-            "the cache-hit path must issue a standalone source option read; commands=${session.execCommands}",
-            session.execCommands.any {
-                it.contains("show-options -v") &&
-                    it.contains("@ps_agent_source") &&
-                    !it.contains("@@PS_RECORDED_KIND@@")
-            },
-        )
-    }
 
     @Test
     fun cachedRecordedKindWithRawSourceRereadsAfterGenerationAppearsAndRebinds() =
